DDMORE-1529: Cat cov vars are now parsed as factors by readDataObj().
Now checks the "use" attribute value of each variable in
DATA_INPUT_VARIABLES and if contains "catCov", treat this variable as a
factor. Factor levels and labels are parsed from the "withCategories..."
part of that "use" attribute value string.

This is a bit clunky since TEL shouldn't have to parse this string
itself; it is an interim solution pending the MDL->JSON->R converter
being updated to properly parse atttibute strings such as this, into R
structures.